requirements.txt
transformers
pillow
torch
requests
matplotlib
streamlit
pylint
